Allow local and remote branches to have different names.
For example, when commiting to local branch "builds/staging" which
tracks branch "master" on remote "staging", the 'git-push' line
silently pushes "builds/staging" to "staging/builds/staging" despite
one's local git config for this branch.

Doing a manual 'git push staging builes/staging' errors and exits
with this message:

```
fatal: The upstream branch of your current branch does not match
the name of your current branch.  To push to the upstream branch
on the remote, use

    git push staging HEAD:master

To push to the branch of the same name on the remote, use

    git push staging builds/staging
```

In this scenario, we want to follow the first suggestion, rather
than the latter. This syntax is also valid when the local and upstream
branches' names match.